+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggcube (development version)
 
+* plotmath is now supported in axis text and titles; e.g., `xlab(expression(italic(alpha[1])))` now behaves as expected.
+
 # ggcube 0.1.0
 
 * Initial CRAN submission.